Fixing router sink issues
Router sink receives the input as String through byteArrayTextToString|routerSinkConsumer.
Spring Cloud Stream used to convert the string back to byte[], but the commit below removed that logic.
https://github.com/spring-cloud/spring-cloud-stream/commit/5d9de8ad579d3464d1503d1a5d1390168bccbdb9

Therefore we are adding it back in the router sink app by programmatically converting the String back to
byte[] before sending it out to the bound router channel.