Remove auto-config for `ComponentCustomizer`
Since we still need to modify every single function, consumer & supplier
for previously introduced `@CustomizationAware`, it does not do too much
difference for us if we add a `ComponentCustomizer` dependency to the
bean for customization and call it manually.
There is not going to be a difference for end-user to declare a
`ComponentCustomizer` bean, unless they don't need to worry about
its `static` nature anymore.

* Remove `ComponentCustomizerBeanPostProcessor` and
 respective `ComponentCustomizationAutoConfiguration`
* Use `@Nullable ComponentCustomizer<FileInboundChannelAdapterSpec>`
as dependency in the `FileSupplierConfiguration.fileMessageSource()` bean
* Rename `@CustomizationAware` to `@CustomizationDoc` since we still need
to generate docs for the customizer end-user can add for this or that function
application.

So, may benefit of this change is to avoid a `static` confusion and performance
overhead with extra `BeanPostProcessor`