Remove maven project level settings.xml
A maven extension to allow project level settings.xml
was previously added to fix an issue on Jenkins CI
in #9fbec999552fa1348460c806da65b6b74337e93f

However, this is now causing issues when importing
project into IDEA. Also, main will not be built on
Jenkins so we are removing the extension.

when loading into IDEs
onis no longer needed and was causing
issues when loading the project into IDEA.

This removes the extension.