GH-348: Resolve knownHostsExpression to Resource
Fixes https://github.com/spring-cloud/stream-applications/issues/348

Currently, SFTP modules use a `FileSystemResource` based on a resolved
from knownHostsExpression.
This is not always true since the resource could be anywhere, not only in
the target file system.
The typical use-use is classpath or Secretes on the target platform

* Use `DefaultResourceLoader` to resolve a location from the expression
to any possible `Resource` impl