Fix dependency mgmt in generated apps pom.xml
The previous commit for adding Tracing attempted to use dependency
management for the `prometheus-rsocket-spring` lib in the
generated applications pom.xml. However, the app generator plugin
assumes the dep mgmt is for a POM and adds tags for this which
fails at runtime.
This commit removes the dep. mgmt. for `prometheus-rsocket-spring` and
instead specifies the version on the dependency.

Also, the tracing property for Wavefront and Zipkin are updated.