Update to Spring Boot 3.4.1
This commit updates the following dependencies:

- Spring Boot to 3.4.2-SNAPSHOT
- Spring Cloud 2024.0.1-SNAPSHOT
- Spring Functions Catalog 5.1.0-SNAPSHOT

Also removes these unnecessary version properties:
- spring-cloud-stream.version
- spring-cloud-function.version
- spring-cloud-starters.version

Also removes dependency mgmt for the following:
- testcontainers-bom
- junit-bom
- jakarta-jms-api
- groovy-bom

Resolves #599 #600

* Remove Debezium Avro integration tests

This commit removes the DebeziumSupplierAvroFormatTest as it
relies on a dependency that has been removed in Spring Integration
and stream-applications. We should not need to test Debezium
support for Avro encoding.